Add api route start-unit
