+++ db/models/Lesson.js
@@ -4,10 +4,10 @@ const { Schema } = mongoose;
 
 const lessonSchema = new Schema({
   startTime: { type: Number, required: true },
-  endTime: { type: Number, required: true },
-  jitsiID: { type: Number, required: true },
+  endTime: { type: Number },
+  jitsiId: { type: String, required: true },
 });
 
-const Lesson = mongoose.models.Lesson || mongoose.model("Course", lessonSchema);
+const Lesson = mongoose.models.Lesson || mongoose.model("Lesson", lessonSchema);
 
 export default Lesson;
