--- src/pages/api/[id]/startUnit.js
@@ -1,15 +1,0 @@
-import dbConnect from "db/connect";
-import Lesson from "db/models/Lesson";
-
-export default async function handler(request, response) {
-  await dbConnect();
-  const { id } = request.query;
-
-  switch (request.method) {
-    case "GET":
-      const lesson = await Lesson.findById(id);
-      response.status(200).json(lesson);
-      default:
-        return response.status(404).json({ status: "Not found" });
-  }
-}
